refactor(restructure): :sparkles: Flattened project structure. Started changing config for better management through project.
To have a means of creating apps quickly that are also highly configurable I decided to start adding code to to manage-fastapi.
