+++ manage_fastapi/templates/project/{{ cookiecutter.folder_name }}/app/main.py
@@ -1,12 +1,12 @@
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
-from app.core.config import settings
-
+from .config import settings
+from .routes import auth
 
 def get_application():
     _app = FastAPI(title=settings.PROJECT_NAME)
-
+    _app.include_router(auth.router)
     _app.add_middleware(
         CORSMiddleware,
         allow_origins=[str(origin) for origin in settings.BACKEND_CORS_ORIGINS],
