+++ manage_fastapi/templates/project/{{ cookiecutter.folder_name }}/tests/test_bearer.py
