--- potato/.pre-commit-config.yaml
@@ -1,36 +1,0 @@
-repos:
-  - repo: https://github.com/myint/autoflake
-    rev: v1.4
-    hooks:
-      - id: autoflake
-        exclude: .*/__init__.py
-        args:
-          - --in-place
-          - --remove-all-unused-imports
-          - --expand-star-imports
-          - --remove-duplicate-keys
-          - --remove-unused-variables
-  - repo: local
-    hooks:
-      - id: flake8
-        name: flake8
-        entry: flake8
-        language: system
-        types: [python]
-  - repo: https://github.com/pre-commit/mirrors-isort
-    rev: v5.4.2
-    hooks:
-      - id: isort
-        args: ["--profile", "black"]
-  - repo: local
-    hooks:
-      - id: mypy
-        name: mypy
-        entry: mypy
-        language: system
-        types: [python]
-  - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v3.3.0
-    hooks:
-      - id: trailing-whitespace
-      - id: end-of-file-fixer
