--- potato/README.md
@@ -1,7 +1,0 @@
-# potato
-
-This project was generated via [manage-fastapi](https://github.com/ycd/manage-fastapi)! :tada:
-
-## License
-
-This project is licensed under the terms of the MIT license.
