--- potato/app/__init__.py
@@ -1,1 +1,0 @@
-from loguru import logger
