--- potato/app/database.py
@@ -1,16 +1,0 @@
-from sqlalchemy import create_engine
-from sqlalchemy.ext.declarative import as_declarative, declared_attr
-from sqlalchemy.orm import sessionmaker
-
-from .config import settings
-
-engine = create_engine(settings.DATABASE_URI, pool_pre_ping=True)
-SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
-
-
-@as_declarative()
-class Base:
-
-    @declared_attr
-    def __tablename__(cls) -> str:
-        return cls.__name__.lower()
