--- potato/docker-compose.yaml
@@ -1,29 +1,0 @@
-version: "3.8"
-
-services:
-  app:
-    build: .
-    env_file:
-      - .env
-    ports:
-      - "8000:8000"
-    depends_on:
-      - redis
-      - database
-  
-  redis:
-    image: redis:alpine
-    ports:
-      - "6379:6379"
-    volumes:
-      - ./redis.conf:/usr/local/etc/redis/redis.conf
-    restart: always
-
-
-  database:
-    image: postgres:12
-    env_file:
-      - .env
-    ports:
-      - "5432:5432"
-
