--- potato/pyproject.toml
@@ -1,27 +1,0 @@
-[tool.poetry]
-name = "potato"
-version = "0.1.0"
-description = ""
-authors = ["kivo360 <kivo360@gmail.com>"]
-
-[tool.poetry.dependencies]
-python = "^3.9"
-fastapi = "^0.70.1"
-dynaconf = {extras = ["vault", "redis"], version = "^3.1.7"}
-loguru = "^0.5.3"
-uvicorn = "^0.16.0"
-pydantic = {extras = ["dotenv", "email"], version = "^1.9.0"}
-
-[tool.poetry.dev-dependencies]
-pytest = "^5.2"
-pytest-cov = "^2.10.1"
-autoflake = "^1.4"
-flake8 = "^3.8.4"
-mypy = "^0.790"
-isort = "^5.0"
-pre-commit = "^2.8.2"
-black = "^20.8b1"
-
-[build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
