--- potato/setup.cfg
@@ -1,17 +1,0 @@
-[isort]
-profile = black
-known_first_party = app
-
-[flake8]
-max-complexity = 7
-statistics = True
-max-line-length = 88
-ignore = W503,E203
-per-file-ignores =
-    __init__.py: F401
-
-[mypy]
-plugins = pydantic.mypy
-ignore_missing_imports = True
-follow_imports = skip
-strict_optional = True
