+++ pyproject.toml
@@ -32,6 +32,7 @@ pydantic = {extras = ["email"], version = "^1.7.2"}
 python = "^3.6.1"
 questionary = "^1.10.0"
 typer = "^0.4.0"
+shellingham = "^1.4.0"
 
 [tool.poetry.dev-dependencies]
 autoflake = "^1.4"
